+++ templates/manage_appointment.html
@@ -34,7 +34,7 @@
             <td>{{ appointment[3] }}</td>  <!-- ApptTime -->
             <td>{{ appointment[4] }}</td>  <!-- AppStatus -->
             <td>{{ appointment[5] }}</td>  <!-- AppReason -->
-            <td><a href="{{ url_for('view_patient', patient_id=appointment[1], appt_id=appointment[0]) }}" class="btn btn-primary">View
+            <td><a href="{{ url_for('staff.view_patient', patient_id=appointment[1], appt_id=appointment[0]) }}" class="btn btn-primary">View
                 Patient</a></td>
         </tr>
         {% endfor %}
